backend/requirements/requirements.txt
Django==3.2.19
pytz==2023.3
sqlparse==0.4.4
requests==2.26.0
djangorestframework==3.12.4
pillow==9.5.0
djoser==2.1.0
django-filter==23.2
drf-extra-fields==3.5.0
filetype==1.2.0
python-dotenv==0.21.1
psycopg2-binary==2.9.6
gunicorn==20.0.4
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.8.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.5.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.10.1
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.5.6
sqlparse==0.4.4
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.26.20